requirements.txt
streamlit
torch
transformers
accelerate
pillow
numpy
matplotlib
pandas
torch
torchvision
transformers
datasets
Pillow
numpy
tqdm
pycocoevalcap
streamlit
matplotlib
pandas
scikit-learn